requirements.txt
aerich==0.5.3
aiofiles==0.5.0
aiosqlite==0.16.1
amqp==5.0.6
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.22.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
ddlparse==1.9.0
Deprecated==1.2.12
dictdiffer==0.8.1
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
fastapi-login==1.6.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httpcore==0.13.0
httptools==0.1.2
httpx @ https://github.com/encode/httpx/archive/refs/tags/0.18.0.tar.gz
idna==2.10
importlib-metadata==4.0.1
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
makefun==1.11.3
MarkupSafe==1.1.1
orjson==3.5.2
passlib==1.7.4
promise==2.3
prompt-toolkit==3.0.18
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
pyparsing==2.4.7
pypika-tortoise==0.1.0
python-dotenv==0.17.0
python-multipart==0.0.5
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
Rx==1.6.1
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
tortoise-orm==0.17.2
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
wrapt==1.12.1
zipp==3.4.1
